requirements.txt
edge-tts==6.1.12
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
pytest==8.3.5
